refactor: extract shared channel grouping, trim comments
Hoist the duplicated filter-and-group loop (navigator, channels overview, and
the Post sheet's channel picker) into a single groupChannels helper in
lib/board. Drop the redundant existingChannel fallback and a few comments that
restated the adjacent code.
